feat: refactor layouts and components for improved structure and navigation - Replace PlatformLayout imports across multiple files - Remove unused ArticleLayout and XmlLayout components - Introduce DocsLayout and LegalLayout for documentation and legal pages - Enhance SideLayout with additional card components for better UI - Implement Article component for shared article content rendering
