+++ web/src/components/Auth.tsx
@@ -2,8 +2,8 @@ import type { ReactElement } from 'react';
 import { useLocation } from 'react-router';
 import { Center } from '@astryxdesign/core/Center';
 import NotFound from '@/pages/NotFound';
+import PlatformLayout from '@/pages/layout';
 import { useUserProfile } from '@/hooks/use-user';
-import PlatformLayout from '@/layout/PlatformLayout';
 import { SignInCard } from '@/components/SignInCard';
 import { hasMinimumRole, type PlatformRole } from '@/lib/roles';
