+++ web/src/pages/Settings.tsx
@@ -17,8 +17,8 @@ import { EmptyState } from '@astryxdesign/core/EmptyState';
 import { Building2, Paintbrush, Settings2, UserRound } from 'lucide-react';
 import { SideNav, SideNavItem, SideNavSection } from '@astryxdesign/core/SideNav';
 import { Table, type TableColumn, pixel, proportional } from '@astryxdesign/core/Table';
+import PlatformLayout from '@/pages/layout';
 import { useDeleteDialog } from '@/lib/utils';
-import PlatformLayout from '@/layout/PlatformLayout';
 import { PageContainer } from '@/components/PageContainer';
 import { useDeleteOrganization } from '@/hooks/use-organization';
 import CreateOrganization from '@/components/dialogs/CreateOrganization';
